V11.2 (#40)
* Updates the version to 11.2-SNAPSHOT

* Adds the new version and updates the release date

* Adds notes on the creation of a new branch

* Updates the ProductSummary to ProductMetadata

* Updates to the MPD code

* Updates the xsd hybrid file to be closer to the v003

* Updates the xsd hybrid file to be closer to v003

* Updates the XML files to follow the latest v003 XSD & updates to the code to respect the changes

* Updates the activeMQ version

* Updates the xalan version

* Updates the ant project version

* Small update of if condition

* Improvements to the MPD services

* Renames the UC1 EX1 Test file

* Adds test 5 and test 6

* Beautification of the test code

* Beautification of the test code

* Adds a simple static method to create a Time object for a certain date

* Adds to UC1 Ex1 the Test Cases 7, 8, and 9

* Adds one more log message

* Removes unused import

* Removes duplication of code

* Approximates the hybrid xsd, to the v003

* Approximates the hybrid xsd to the v003

* Renames the type to token

* Improves the generation of the docx files

* Improves the text of the MPD specification

* Updates the code to generate the Type Signature details together with the name of the field

* Improves the MPD specificatio

* Adds Type Signature Details as bullets

* Appends the operation number to be part of the key of the Publisher sources

* Updates the hour to 12h (middle of the day)

* Fixes the numbering issue with MO Errors section

* Updates the xml file of the MPD services

* Updates the description of the source fields

* Batch updates to MPD

* Batch updates to MPD (2)

* Changes the ProductType from an MO Object to a MO Composite

* Changes the prefix for MO Objects on the book auto-generation

* Reshuffles the Product and ProductType in the xml file

* Changes the orderID from Identifier to Long

* Renames the timeWindow to contentDate

* Updates the numerical values of the SecondaryErrorCodeEnum

* Improves the text of the specification

* Updates de description of the orderID field

* Fixes the Bookmarks bug

* Removes the use of the word 'parameter' and replaces it with 'attribute'

* Fixes the description to match the name of the service

* Removes the use of the word 'parameter' and replaces it with 'attribute' (2)

* Updated the MO Errors of some of the operations

* Small fix - Integer is in MAL not MPD

* Cleans up the Data structures text of MPD

* Updates the text

* General improvements on the requirements

* The oftware can now render the NOTE correctly in the .docx files

* Optimizes the code for the NOTEs

* Capitalizes the first word of the NOTEs

* Abstracting the MPD Tests

* Further abstraction

* Moves commmon code to the abstract class

* Renames the file from MPSTest to MPDTest

* Starts the UC3_Ex1_Test file

* Batch of code for Test 1 (not completed)

* Adds a Listener for when new Products become available

* Small MPD code improvements

* Small MPD code improvements (2)

* Updates the Backend to use the Dataset class

* Updates the yaml file to have the http on jdk11

* Adds the http dependency from Sun

* Restore to http testbed build to Java 8

* General Updates for MPD

* Add also the subscription keys fields in the Type Signature Description section

* Improvements on the text and removal of unused Functional Requirements sections

* Adds 4 more tests to UC3_Ex1

* Adds 2 more tests to UC3_Ex1 for testing with different users

* Adds more tests to UC3_Ex1 for testing with different domains

* Adds 1 more test to UC3_Ex1 for testing with the correct domains

* Code cleanup

* Introduces the SERVICE_JUST_METADATA delivery method option

* Updates the Release Notes with the changes done for version 11.2

* Updates the javadocs

* Updates versions to 11.2

* Updates the actions from v3 to v4

* Updates the report names to include the jdk version being used

+++ api-generator/generator-java/src/main/java/esa/mo/tools/stubgen/GeneratorLangs.java
@@ -1639,7 +1639,8 @@ public abstract class GeneratorLangs extends GeneratorBase {
         createListClass(folder, area, service, className, abstractComposite, composite.getShortFormPart());
     }
 
-    public abstract void createListClass(File folder, AreaType area, ServiceType service, String srcTypeName, boolean isAbstract, Long shortFormPart) throws IOException;
+    public abstract void createListClass(File folder, AreaType area, ServiceType service,
+            String srcTypeName, boolean isAbstract, Integer shortFormPart) throws IOException;
 
     protected final void createMultiReturnType(String destinationFolderName, String returnTypeFqName, MultiReturnType returnTypeInfo) throws IOException {
         logger.info(" > Creating multiple return class class " + returnTypeFqName);
